alguns ajustes que esqueci de commitar rs
