+++ db/mongodb.go
@@ -232,14 +232,17 @@ func (m *MongoDB) ExtraIndexes(idxs []string) error {
 		e := strings.Split(v, ".")
 		if len(e) > 1 {
 			v = strings.ReplaceAll(e[1], "_", "")
-			if strings.Contains(e[0], "qsa") {    // Se for qsa.index1 ele so converte para qsa_index1
+			if e[0] == "qsa" {
 				v = fmt.Sprintf("%s.%s", "qsa", v)
 			}
-			if strings.Contains(e[0], "secundario") && strings.Contains(e[0], "cnae") { //Aqui no json esta cnaes_secundarios se o usuario digitar qualquer coisa proxima ele converte para o padrao
+			if e[0] == "cnaes_secundarios" {
 				v = fmt.Sprintf("cnaesecundarios.%s", e[1])
+			} else {
+				log.Output(1, fmt.Sprintf("the index %s does not exist in json, so it cannot be created.", e[0]))
+				return nil
 			}
 		}
-		v = fmt.Sprintf("json.%s", v) // concatena o json. que e o nome da raiz
+		v = fmt.Sprintf("json.%s", v)
 		i = append(i, mongo.IndexModel{
 			Keys: bson.D{{Key: v, Value: 1}},
 		})
@@ -251,4 +254,28 @@ func (m *MongoDB) ExtraIndexes(idxs []string) error {
 	log.Output(1, fmt.Sprintf("%d indexes successfully created in the collection %s", len(r), companyTableName))
 	return nil
 
+}
+
+func (m *MongoDB) checkIndexes(indexes []string) (map[string]bool, error) {
+	c := m.db.Collection(companyTableName)
+	cur, err := c.Indexes().List(m.ctx)
+	if err != nil {
+		return nil, err
+	}
+	defer cur.Close(m.ctx)
+	idxs := make(map[string]bool)
+	for cur.Next(m.ctx) {
+		var idx bson.M
+		if err := cur.Decode(&idx); err != nil {
+			return nil, err
+		}
+		if n, ok := idx["name"].(string); ok {
+			idxs[n] = true
+		}
+	}
+	r := make(map[string]bool)
+	for _, index := range indexes {
+		r[index] = idxs[index]
+	}
+	return r, nil
 }
